:sparkles: 4.21*: bloglist expansion, step9
Secure delete route

+++ controllers/blogs.js
@@ -50,6 +50,18 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
